train ZnxZnxC4 w TwoLayerMLP, QuadraticRNN

+++ src/groups/irrep.py
@@ -16,7 +16,7 @@ class IrreducibleRepresentation:
         if matrices.ndim != 3 or matrices.shape[1] != matrices.shape[2]:
             raise ValueError(f"matrices must have shape (n_elements, d, d), got {matrices.shape}")
         self._name = name
-        self._matrices = np.asarray(matrices, dtype=np.float64)
+        self._matrices = np.asarray(matrices)
         self._size = int(matrices.shape[1])
 
     @property
